Implementación de Query para RecomendacionDieta. Se corrigió los services de Dieta y Mascota y se mantiene query para Dieta

+++ src/main/java/com/upc/petminder/dtos/RecordatoriosDTO/ContarRecordatoriosxMascotaDto.java
@@ -8,5 +8,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ContarRecordatoriosxMascotaDto {
-    private int total_recordatorios;
+    private Long total_recordatorios;
 }
